Split StandardTable stories into multiple files and story folders.
